chore: move twitter commands
